compat: fixed access to status attributes
as of Mercurial c5548b0b6847, status is no more a tuple. Its attributes
must be accessed explicitely.

These attributes have been available since 3.2 (introduced in cb4449921a1d),
so that we don't need to introduce a conditional such as `isinstance(tuple)`.

Technically, these are slots, so that performance-wise, such prefetching in
locals should not matter, but they make for the smallest changeset.